refactor: isolate Codex agent support
