+++ lib/hive/agent_profiles/codex.rb
@@ -1,5 +1,4 @@
 require "hive/agent_profile"
-require "hive/skill_check"
 
 module Hive
   module AgentProfiles
@@ -8,10 +7,8 @@ module Hive
       skill_syntax_format: "/%{skill}",
       status_detection_mode: :output_file_exists,
       billing_semantics: :subscription_backed,
-      skill_verifier: Hive::SkillCheck::Codex.method(:verify),
-      default_model_resolver: ->(**kwargs) {
-        Hive::ImplementationIdentity::NativeDefaults.resolve(:codex, **kwargs)
-      },
+      skill_verifier: Hive::AgentSupport.skill_verifier(:codex),
+      default_model_resolver: Hive::AgentSupport.model_resolver(:codex),
       routed_model_argument_builder: ->(model) {
         %w[default inherit].include?(model) ? [] : [ "--model", model ]
       },
