+++ lib/hive/agent_profiles/pi.rb
@@ -9,12 +9,8 @@ module Hive
       skill_syntax_format: "/skill:%{skill}",
       status_detection_mode: :output_file_exists,
       billing_semantics: :subscription_backed,
-      skill_verifier: ->(invocation, project_root: nil) {
-        Hive::AgentSupport.for(PI)::Skills.verify(invocation, project_root: project_root)
-      },
-      default_model_resolver: ->(**kwargs) {
-        Hive::AgentSupport.for(PI).default_model(**kwargs)
-      },
+      skill_verifier: Hive::AgentSupport.skill_verifier(:pi),
+      default_model_resolver: Hive::AgentSupport.model_resolver(:pi),
       routed_model_argument_builder: ->(model) {
         %w[default inherit].include?(model) ? [] : [ "--model", model ]
       }
