+++ lib/hive/commands/doctor.rb
@@ -6,9 +6,6 @@ require "hive"
 require "hive/config"
 require "hive/invoked_binary"
 require "hive/agent_profiles"
-require "hive/agent_profiles/claude"
-require "hive/agent_profiles/codex"
-require "hive/agent_profiles/pi"
 require "hive/claude_launcher"
 require "hive/agent_skills"
 require "hive/web/environment"
