+++ lib/hive/implementation_identity.rb
@@ -246,8 +246,6 @@ module Hive
           case provider.to_sym
         when :claude
           json_model(paths(project_root, home, ".claude/settings.json"), %w[model])
-        when :codex
-          toml_model(paths(project_root, home, ".codex/config.toml"))
         when :grok
           json_model(paths(project_root, home, ".grok/settings.json"), %w[model defaultModel]) ||
             toml_model(paths(project_root, home, ".grok/config.toml"))
