+++ lib/hive/stages/artifacts.rb
@@ -419,9 +419,7 @@ module Hive
           [ task.folder, task.worktree_path ].compact.uniq
         end
         result = nil
-        permission_arguments = if role == "producer" && profile.name == :codex
-          producer_permission_arguments
-        end
+        permission_arguments = producer_permission_arguments if role == "producer"
         begin
           result = Hive::Stages::Base.spawn_agent(
             task,
@@ -609,18 +607,13 @@ module Hive
         )
         required = Array(claims).map { |claim| claim.fetch("proof_kind") }.uniq
         unsupported = required - [ "document" ]
-        if unsupported.any? && !%i[codex pi].include?(profile.name)
+        support = Hive::AgentSupport.for(profile)
+        if unsupported.any? && !support&.respond_to?(:validate_capture_profile!)
           raise Hive::ConfigError,
                 "artifacts evidence producer #{profile.name.inspect} cannot use the managed " \
                 "agent-browser capture boundary; configure a Codex or Pi producer"
         end
-        if unsupported.any? && profile.name == :codex &&
-           (!profile.workspace_write_supported? || !profile.add_dir_flag)
-          raise Hive::ConfigError,
-                "artifacts evidence producer #{profile.name.inspect} cannot safely produce " \
-                "#{unsupported.join(', ')} proof; configure a workspace-sandboxed producer " \
-                "with per-attempt writable roots"
-        end
+        support.validate_capture_profile!(profile:, unsupported:) if unsupported.any?
         profile
       end
