+++ wiki/modules/agent_support.md
@@ -4,17 +4,19 @@ type: module
 source: lib/hive/agent_support.rb, lib/hive/agent_support/
 created: 2026-08-25
 updated: 2026-08-25
-tags: [agent, provider, boundary, selective-loading, pi, opencode]
+tags: [agent, provider, boundary, selective-loading, pi, opencode, codex]
 ---
 
 **TLDR**: `Hive::AgentSupport` is a small convention loader for optional
 built-in provider behavior. It has no registry objects, policy hierarchy,
-cache, or process host. Selecting Pi or OpenCode loads only that provider;
+cache, or process host. Selecting Pi, OpenCode, or Codex loads only that provider;
 selecting an unmigrated provider returns `nil` and preserves its current path.
 
 ## Boundary
 
-The data-only `BUILTINS` map is the complete discovery surface. A provider
+The data-only `BUILTINS` map is the complete discovery surface, and the small
+`DEFAULT_PROMPT_STYLES` table preserves custom-profile CLI compatibility
+without loading a provider. A provider
 root declares cohesive behavior and lazily exposes larger facets by constant:
 
 - `Runtime` compiles provider-specific managed policy while the generic runtime
@@ -27,10 +29,16 @@ root declares cohesive behavior and lazily exposes larger facets by constant:
 - OpenCode additionally owns its typed configuration, bounded run/export
   transaction, launch-scope translation, route observation normalization,
   exact-route/default-model policy, and completion-protocol exception.
+- Codex owns its native reviewer, managed-runtime and evidence permissions,
+  path-qualified reads, skill/plugin inventory, setup, credential, model,
+  effort, and invocation rules. The generic runtime still materializes output
+  schemas; `Hive::Reviewers::Runtime` owns its subprocess and findings writes,
+  and the reviewer receives its generic stage host rather than loading
+  orchestration or process custody into the provider package.
 
 Generic callers resolve the selected support once at their existing seam and
 use the named facet. They do not switch on a migrated provider. Loading the
-profile catalog loads neither Pi nor OpenCode; selecting one loads its small
+profile catalog loads none of Pi, OpenCode, or Codex; selecting one loads its small
 root/configuration, while execution, skills, and setup remain lazy.
 
 Core authority does not move: Hive still starts and reaps processes, writes
