+++ wiki/stages/review.md
@@ -73,7 +73,7 @@ Runs `cfg.review.ci.command` (e.g., `bin/ci`) once on entry. The subprocess is l
 
 ## Phase 2 — reviewers (`Hive::Reviewers`)
 
-For each selected reviewer spec, sequentially: dispatch via `Hive::Reviewers.dispatch(spec, ctx)`, run the selected adapter, and write `reviews/<output_basename>-<NN>.md`. `kind: "agent"` uses `Hive::Reviewers::Agent` and `Hive::Agent.run!` with the spec's profile; `kind: "codex_review"` uses `Hive::Reviewers::CodexReview`, which runs native `codex review --title <title> <prompt>` and captures valid High/Medium/Nit output into the findings file. The native Codex adapter normalizes stdout before publishing it: after the first severity header, it drops the verbose middle `exec` / `thinking` / `codex` session transcript and keeps codex's final assistant message, so triage reads findings rather than hundreds of KB of tool logs. Normal tasks select specs from `cfg.review.reviewers`; synthetic patrol handoff tasks whose `task.md` frontmatter carries `source: patrol` select specs from `cfg.patrol.review.reviewers` instead, so patrol PRs can use the cheap `codex-native-review` default without changing the normal feature-review policy. Ad-hoc PR tasks whose `task.md` frontmatter carries `source: ad-hoc` select `cfg.review.adhoc.reviewers` when that list is set, otherwise they fall back to `cfg.review.reviewers`. Reviewer prompts compare against `origin/<default_branch>` when that remote-tracking ref exists (probed via the full `refs/remotes/origin/<branch>` path so a like-named tag cannot satisfy the check), falling back to the configured/default local branch only when the remote ref is absent; this keeps long-lived local `main` checkouts from making reviewers report already-merged changes as current-task findings. `hive review --pr` records the PR's declared `baseRefName` in `pr.md`, but v1 still uses the existing default-branch compare path rather than a per-task base override. When `cfg.default_branch` is unset AND `Hive::GitOps#origin_default_branch` returns nil (no origin/HEAD symref and neither `origin/main` nor `origin/master` exist), review preflight refuses to fall back to the worktree's current branch (which in a `git worktree add` is the task branch — diffing the task against itself produces zero or phantom findings) and exits 1 with the two remediation paths named: set `default_branch` in `.hive-state/config.yml`, or run `git remote set-head origin --auto` so origin/HEAD points at the project default.
+For each selected reviewer spec, sequentially: dispatch via `Hive::Reviewers.dispatch(spec, ctx)`, run the selected adapter, and write `reviews/<output_basename>-<NN>.md`. `kind: "agent"` uses `Hive::Reviewers::Agent` and `Hive::Agent.run!` with the spec's profile; `kind: "codex_review"` lazily selects `Hive::AgentSupport::Codex::Reviewer`, which runs native `codex review --title <title> <prompt>` and captures valid High/Medium/Nit output into the findings file. The native Codex adapter normalizes stdout before publishing it: after the first severity header, it drops the verbose middle `exec` / `thinking` / `codex` session transcript and keeps codex's final assistant message, so triage reads findings rather than hundreds of KB of tool logs. Normal tasks select specs from `cfg.review.reviewers`; synthetic patrol handoff tasks whose `task.md` frontmatter carries `source: patrol` select specs from `cfg.patrol.review.reviewers` instead, so patrol PRs can use the cheap `codex-native-review` default without changing the normal feature-review policy. Ad-hoc PR tasks whose `task.md` frontmatter carries `source: ad-hoc` select `cfg.review.adhoc.reviewers` when that list is set, otherwise they fall back to `cfg.review.reviewers`. Reviewer prompts compare against `origin/<default_branch>` when that remote-tracking ref exists (probed via the full `refs/remotes/origin/<branch>` path so a like-named tag cannot satisfy the check), falling back to the configured/default local branch only when the remote ref is absent; this keeps long-lived local `main` checkouts from making reviewers report already-merged changes as current-task findings. `hive review --pr` records the PR's declared `baseRefName` in `pr.md`, but v1 still uses the existing default-branch compare path rather than a per-task base override. When `cfg.default_branch` is unset AND `Hive::GitOps#origin_default_branch` returns nil (no origin/HEAD symref and neither `origin/main` nor `origin/master` exist), review preflight refuses to fall back to the worktree's current branch (which in a `git worktree add` is the task branch — diffing the task against itself produces zero or phantom findings) and exits 1 with the two remediation paths named: set `default_branch` in `.hive-state/config.yml`, or run `git remote set-head origin --auto` so origin/HEAD points at the project default.
 
 The patrol selector is deliberately frontmatter-based and fail-soft:
 `reviewer_specs_for` reads YAML only when `task.md` starts with a
